// Educational.Client/Educational.Client/client/src/api/api.ts
export interface TeacherForCourseDetailVm {
    userId?: string;
    name?: string | undefined;
}

export interface UpdateCourseDto {
    id?: string;
    name?: string | undefined;
    description?: string | undefined;
    duration?: number;
    price?: number;
    confirmed?: boolean;
    teacherId?: string;
}

export interface UpdateLessonDto {
    id?: string;
    name?: string | undefined;
    description?: string | undefined;
    number?: number;
}